% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.dictionary.textmodel_wordmap}
\alias{as.dictionary.textmodel_wordmap}
\alias{as.list.textmodel_wordmap}
\title{Extract features with largest coefficients from a wordmap model}
\usage{
\method{as.dictionary}{textmodel_wordmap}(x, separator = NULL, ...)

\method{as.list}{textmodel_wordmap}(x, ...)
}
\arguments{
\item{separator}{the character in between multi-word dictionary values. If
\code{NULL}, \code{x$concatenator} will be used.}

\item{...}{passed to \link{coef.textmodel_wordmap}}
}
\description{
Extract features with largest coefficients from a wordmap model
}
