% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{as.dictionary.textmodel_wordmap}
\alias{as.dictionary.textmodel_wordmap}
\alias{as.list.textmodel_wordmap}
\title{Create lexicon from a Wordmap model}
\usage{
\method{as.dictionary}{textmodel_wordmap}(x, separator = NULL, ...)

\method{as.list}{textmodel_wordmap}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link[=textmodel_wordmap]{textmodel_wordmap()}}.}

\item{separator}{the character in between multi-word dictionary values. If
\code{NULL}, \code{x$concatenator} will be used.}

\item{...}{passed to \link{coef.textmodel_wordmap}}
}
\value{
Returns a list or a \link[quanteda:dictionary]{quanteda::dictionary} object.
}
\description{
\code{as.list()} returns features with the largest coefficientsa as a list of
character vector. \code{as.dictionary()} returns a \link[quanteda:dictionary]{quanteda::dictionary} object
that can be use for dictionary analysis.
}
