% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{coef.textmodel_wordmap}
\alias{coef.textmodel_wordmap}
\alias{coefficients.textmodel_wordmap}
\title{Extract coefficients for features}
\usage{
\method{coef}{textmodel_wordmap}(object, n = 10, select = NULL, ...)

\method{coefficients}{textmodel_wordmap}(object, n = 10, select = NULL, ...)
}
\arguments{
\item{object}{a Newsmap model fitted by \code{\link[=textmodel_wordmap]{textmodel_wordmap()}}.}

\item{n}{the number of coefficients to extract.}

\item{select}{returns the coefficients for the selected class; specify by the
names of rows in \code{object$model}.}

\item{...}{not used.}
}
\description{
Extract coefficients for features
}
