% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_topic}
\alias{data_dictionary_topic}
\title{Seed topic dictionary}
\description{
A dictionary with seed words for size common topics at the United Nations
General Assembly.
}
\references{
Watanabe, Kohei & Zhou, Yuan (2020). "Theory-Driven Analysis of
Large Corpora: Semisupervised Topic Classification of the UN Speeches".
doi:10.1177/0894439320907027. \emph{Social Science Computer Review}.
}
\author{
Kohei Watanabe \email{watanabe.kohei@gmail.com}
}
\keyword{data}
